+++ lib/pragma/policy.rb
@@ -1,15 +1,13 @@
 # frozen_string_literal: true
 
-require 'pundit'
-
 require 'pragma/policy/version'
+require 'pragma/policy/scope'
 require 'pragma/policy/base'
+require 'pragma/policy/pundit'
 require 'pragma/policy/errors'
 
 module Pragma
   # Fine-grained access control for your API resources.
-  #
-  # @author Alessandro Desantis
   module Policy
   end
 end