+++ lib/pragma/policy/base.rb
@@ -7,40 +7,9 @@ module Pragma
     # A policy provides predicate methods for determining whether a user can perform a specific
     # action on a record.
     #
-    # @author Alessandro Desantis
-    #
     # @abstract Subclass and implement action methods to create a policy.
     class Base
-      # Authorizes AR scopes and other relations by only returning the records accessible by the
-      # current user. Used, for instance, in index operations.
-      #
-      # @author Alessandro Desantis
-      class Scope
-        # @!attribute [r] user
-        #   @return [Object] the user accessing the records
-        #
-        # @!attribute [r] scope
-        #   @return [Object] the relation to use as a base
-        attr_reader :user, :scope
-
-        # Initializes the scope.
-        #
-        # @param user [Object] the user accessing the records
-        # @param scope [Object] the relation to use as a base
-        def initialize(user, scope)
-          @user = user
-          @scope = scope
-        end
-
-        # Returns the records accessible by the given user.
-        #
-        # @return [Object]
-        #
-        # @abstract Override to implement retrieving the accessible records
-        def resolve
-          fail NotImplementedError
-        end
-      end
+      Scope = ::Pragma::Policy::Scope
 
       # @!attribute [r] user
       #   @return [Object] the user operating on the record