+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.1.0]
+
 ### Added
 
 - Added `Pragma::Policy::Pundit` for policies that just delegate to Pundit
@@ -25,5 +27,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 First Pragma 2 release.
 
-[Unreleased]: https://github.com/pragmarb/pragma-policy/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/pragmarb/pragma-policy/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/pragmarb/pragma-policy/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/pragmarb/pragma-policy/compare/v0.1.0...v2.0.0